[Inductor max autotune] Make autotune_select_algorithm more robust
This diff makes sure that a custom exception is thrown when no valid
choices remain during autotuning. This allows to gracefully fall back
to a default choice, even if that default choice has not been passed to
autotune_select_algorithm.

Additionally, this diff handles RuntimeErrors during autotuning gracefully, e.g. the corresponding choice is ignored but it does not lead to the compilation failure of the entire model if a problematic choice is encountered during autotuning.
( An error is being logged, though).

ghstack-source-id: 904251b11a684054ab1c734a307d5a0b0f910a05
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124928